refactor(platform): split platform/views.py into domain modules
Closes #642
